# [core] 1/N. ray.wait(_, num_returns=1) performance improvements