Improve performance of `ray.wait(_, num_returns=1)` by early exiting the search for owned objects.

Before
```Python
In [9]: @ray.remote
   ...: def f(i):
   ...:     return str(i)


In [9]: def g(rest, n_ret):
   ...:     while True:
   ...:         done, rest = ray.wait(rest, num_returns=n_ret)
   ...:         if (len(rest) == 0):
   ...:             return
   ...:         ray.get(done)


In [9]: objs = [ f.remote(i) for i in range(10000) ]

// Before
In [8]: %time g(objs, 1)
CPU times: user 15.1 s, sys: 120 ms, total: 15.2 s
Wall time: 15.1 s

// After
In [7]: %time g(objs, 1)
CPU times: user 15.5 s, sys: 1.89 s, total: 17.4 s
Wall time: 13.5 s
```

CPU Time Before (20% time spend in ReferenceCounter::HasOwner)
![head_100k_perf](https://github.com/user-attachments/assets/6384cc1b-38d6-46a4-812c-c2cdb782611a)

CPU Time After 
![head_100k_early-exit](https://github.com/user-attachments/assets/b70ba243-d7a9-417b-ac2a-bf8eabc339a0)
(ReferenceCounter::HasOwner has no overhead)

There are still more improvements to be made. See #49905 for more details.